first revision of afni4d support

+++ neurovault/apps/statmaps/forms.py
@@ -18,9 +18,8 @@ from .models import Collection, Image
 from django.forms.forms import Form
 from django.forms.fields import FileField
 import tempfile
-from neurovault.apps.statmaps.utils import split_filename, get_paper_properties
-from django.core.files.base import File, ContentFile
-from django.core.files.uploadedfile import InMemoryUploadedFile
+from neurovault.apps.statmaps.utils import split_filename, get_paper_properties, \
+                                        detect_afni4D, split_afni4D_to_nii, memory_uploadfile
 from django import forms
 
 # Create the form class.
@@ -217,7 +216,7 @@ collection_row_attrs = {
 
 
 class CollectionForm(ModelForm):
-    
+
     class Meta:
         exclude = ('owner','private_token')
         model = Collection
@@ -273,12 +272,15 @@ class ImageForm(ModelForm):
         self.helper = FormHelper(self)
         self.helper.form_class = 'form-horizontal'
         self.helper.form_tag = False
+        self.afni_briks = []
+        self.afni_tmp = None
 
     class Meta:
         model = Image
         exclude = ('json_path', 'collection')
 
-    def clean(self):
+    def clean(self, **kwargs):
+
         cleaned_data = super(ImageForm, self).clean()
         file = cleaned_data.get("file")
 
@@ -314,13 +316,14 @@ class ImageForm(ModelForm):
                         return cleaned_data
 
                 # write the data file to a temporary directory
-                f = open(os.path.join(tmp_dir, fname + ext), "wb")
+                nii_tmp = os.path.join(tmp_dir, fname + ext)
+                f = open(nii_tmp, "wb")
                 f.write(file.file.read())
                 f.close()
 
                 # check if it is really nifti
                 try:
-                    nii = nb.load(os.path.join(tmp_dir, fname + ext))
+                    nii = nb.load(nii_tmp)
                 except Exception as e:
                     self._errors["file"] = self.error_class([str(e)])
                     del cleaned_data["file"]
@@ -328,23 +331,56 @@ class ImageForm(ModelForm):
 
                 # convert to nii.gz if needed
                 if ext.lower() != ".nii.gz":
-                    #Papaya does not handle flaot64, but by converting files we loose precision
-#                     if nii.get_data_dtype() == np.float64:
-#                         nii.set_data_dtype(np.float32)
-                    nb.save(nii, os.path.join(tmp_dir, fname + ".nii.gz"))
-                    f = ContentFile(open(os.path.join(tmp_dir, fname + ".nii.gz")).read())
-                    print cleaned_data["file"].__class__.__name__
-                    cleaned_data["file"] = InMemoryUploadedFile(f, "file", fname + ".nii.gz",
-                                                                cleaned_data["file"].content_type, f.size, cleaned_data["file"].charset)
+
+                    #Papaya does not handle float64, but by converting files we loose precision
+                    #if nii.get_data_dtype() == np.float64:
+                    #ii.set_data_dtype(np.float32)
+                    nii_tmp = os.path.join(tmp_dir, fname + ".nii.gz")
+                    nb.save(nii, nii_tmp)
+
+                    cleaned_data['file'] = memory_uploadfile(nii_tmp, fname,
+                                                             cleaned_data['file'])
+
+                # detect AFNI 4D files and prepare 3D slices
+                if nii_tmp is not None and detect_afni4D(nii_tmp):
+                    self.afni_briks = split_afni4D_to_nii(nii_tmp)
+
             finally:
                 try:
-                    shutil.rmtree(tmp_dir)  # delete directory
+                    if self.afni_briks:
+                        self.afni_tmp = tmp_dir  # keep temp dir for AFNI slicing
+                    else:
+                        shutil.rmtree(tmp_dir)
                 except OSError as exc:
                     if exc.errno != 2:  # code 2 - no such file or directory
                         raise  # re-raise exception
         else:
             raise ValidationError("Couldn't read uploaded file")
         return cleaned_data
+
+    def save(self, commit=True):
+        image = super(ImageForm, self).save(commit=False)
+        basenm = image.name
+        if self.afni_briks:
+            #import ipdb;ipdb.set_trace()
+            try:
+                label,brik = self.afni_briks.pop(0)
+                mfile = memory_uploadfile(brik, label, image.file)
+                image.file = mfile
+                image.name = '%s - %s' % (basenm, label)
+                image.save()
+
+                for label,brik in self.afni_briks:
+                    mfile = memory_uploadfile(brik, label, image.file)
+                    brik_img = Image(name='%s - %s' % (basenm, label), file=mfile,
+                                     collection=self.instance.collection)
+                    for field in ['description','map_type','tags']:
+                        setattr(brik_img, field, getattr(image,field))
+                    brik_img.save()
+            finally:
+                shutil.rmtree(self.afni_tmp)
+        else:
+            image.save()
 
 
 class SingleImageForm(ImageForm):
